pyproject.toml
[tool.poetry]
name = "sassbot-beta"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
wheel = "^0.35.1"
ffmpeg = "^1.4"
youtube_dl = "^2020.11.1"
pafy = "^0.5.5"
PyNaCl = "^1.4.0"
requests = "^2.24.0"
PyGithub = "^1.53"
flask = "^1.1.2"
ksoftapi = "^0.2.5-alpha.0"
speedtest-cli = "^2.1.2"
tools = "^0.1.9"
xkcd = "^2.4.2"
googletrans = "^3.0.0"
psutil = "^5.7.3"
Pillow = "^8.0.1"
aiopentdb = "^0.5.0"
jokeapi = "^0.2.6"
replit = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
